Cleanup: move unit tests to the file where the functionality is
